Redid error handling. Now allows better flow by not throwing any exceptions for controller logic. Updated client error handling to better accommodate this.

+++ Application/Groups/Leave.cs
@@ -2,6 +2,7 @@ using System;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Core;
 using Application.Errors;
 using Application.Interfaces;
 using MediatR;
@@ -12,12 +13,12 @@ namespace Application.Groups
 {
     public class Leave
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             private readonly IUserAccessor _userAccessor;
@@ -27,31 +28,28 @@ namespace Application.Groups
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var group = await _context.Groups.FindAsync(request.Id);
                 if (group == null)
-                    throw new RestException(HttpStatusCode.NotFound, new { Group = "Not Found" });
+                    return null;
 
                 var user = await _context.Users.SingleOrDefaultAsync(x => x.UserName == _userAccessor.GetCurrentUsername());
                 if (user == null)
-                    throw new RestException(HttpStatusCode.NotFound, new { User = "Not found" });
+                    return null;
 
                 var membership = await _context.UserGroups.SingleOrDefaultAsync(x => x.GroupId == group.Id && x.AppUserId == user.Id);
 
-                if (membership == null)
-                    return Unit.Value;
-
-                if (membership.IsHost)
-                    throw new RestException(HttpStatusCode.BadRequest, new { Membership = "As the host you cannot leave this group." });
+                if (membership == null || membership.IsHost)
+                    return null;
 
                 _context.UserGroups.Remove(membership);
 
                 var success = await _context.SaveChangesAsync() > 0;
 
-                if (success) return Unit.Value;
+                if (!success) return Result<Unit>.Failure("Failed to leave group. Are you the host?");
 
-                throw new Exception("Problem saving changes ");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
